datafeeder: refactor service.DataUploadValidityService as api.AuthorizationService
