+++ datafeeder/src/test/java/org/georchestra/datafeeder/api/FileUploadApiControllerSecurityTest.java
@@ -22,7 +22,6 @@ import java.util.Collections;
 
 import org.georchestra.datafeeder.service.DataPublishingService;
 import org.georchestra.datafeeder.service.DataUploadService;
-import org.georchestra.datafeeder.service.DataUploadValidityService;
 import org.georchestra.datafeeder.service.FileStorageService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -44,7 +43,7 @@ public class FileUploadApiControllerSecurityTest {
     private @MockBean FileStorageService storageService;
     private @MockBean DataUploadService uploadService;
     private @MockBean ApiResponseMapper mapper;
-    private @MockBean DataUploadValidityService mockDataUploadValidityService;
+    private @MockBean AuthorizationService mockDataUploadValidityService;
     private @MockBean DataPublishingService mockDataPublishingService;
 
     private @Autowired FileUploadApi controller;
